requirements.txt
Django==3.2
cmake==3.17.2
django-filter==2.4.0
django-jchart==0.4.2
django-tex==1.1.10
gunicorn==20.0.4
matplotlib==3.9.2
mysql-connector==2.2.9
mysqlclient==1.4.6
numpy==2.1.1
Pillow==10.4.0
six==1.16.0
sqlparse==0.3.1
python-dateutil
django-simple-captcha
vacances-scolaires-france
django-model-utils==4.0.0
django-utils-six

#Il faut installer django_quizz
